Added settings view

+++ Indulged/Indulged/Plugins/ProCamera/ProCameraPage.xaml.cs
@@ -43,8 +43,13 @@ namespace Indulged.Plugins.ProCamera
         {
             if (executedOnce)
             {
-                NavigationService.GoBack();
-                NavigationService.RemoveBackEntry(); 
+                Dispatcher.BeginInvoke(() => {
+                    if (NavigationService.CanGoBack)
+                    {
+                        NavigationService.GoBack();
+                        NavigationService.RemoveBackEntry();
+                    }
+                });
 
                 return;
             }
@@ -94,8 +99,14 @@ namespace Indulged.Plugins.ProCamera
             }
             else
             {
-                NavigationService.GoBack();
-                NavigationService.RemoveBackEntry();
+                Dispatcher.BeginInvoke(() => {
+                    if (NavigationService.CanGoBack)
+                    {
+                        NavigationService.GoBack();
+                        NavigationService.RemoveBackEntry();
+                    }
+
+                });
             }
         }
